backend/requirements.txt
fastapi
uvicorn
openai
numpy
faiss-cpu
pydantic
deepeval

ragas
datasets
evaluate